Moved the rate limiter to Pipeline
