+++ docetl/runner.py
@@ -11,33 +11,9 @@ from rich.console import Console
 from docetl.dataset import Dataset, create_parsing_tool_map
 from docetl.operations import get_operation
 from docetl.operations.utils import flush_cache
-import pyrate_limiter
-import math
-from inspect import isawaitable
 from .pipeline import Pipeline
 
 load_dotenv()
-
-class BucketCollection(pyrate_limiter.BucketFactory):
-    def __init__(self, **buckets):
-        self.clock = pyrate_limiter.TimeClock()
-        self.buckets = buckets
-
-    def wrap_item(self, name: str, weight: int = 1) -> pyrate_limiter.RateItem:
-        now = self.clock.now()
-
-        async def wrap_async():
-            return pyrate_limiter.RateItem(name, await now, weight=weight)
-
-        def wrap_sync():
-            return pyrate_limiter.RateItem(name, now, weight=weight)
-
-        return wrap_async() if isawaitable(now) else wrap_sync()
-
-    def get(self, item: pyrate_limiter.RateItem) -> pyrate_limiter.AbstractBucket:
-        if item.name not in self.buckets:
-            return self.buckets["unknown"]
-        return self.buckets[item.name]
     
 class DSLRunner(Pipeline):
     """
@@ -87,19 +63,7 @@ class DSLRunner(Pipeline):
             )
 
         self.syntax_check()
-        
-        buckets = {
-            param: pyrate_limiter.InMemoryBucket([
-                pyrate_limiter.Rate(
-                    param_limit["count"],
-                    param_limit["per"] * getattr(pyrate_limiter.Duration, param_limit.get("unit", "SECOND").upper()))
-                for param_limit in param_limits])
-            for param, param_limits in self.config.get("config", {}).get("rate_limits", {}).items()
-        }
-        buckets["unknown"] = pyrate_limiter.InMemoryBucket([pyrate_limiter.Rate(math.inf, 1)])
-        bucket_factory = BucketCollection(**buckets)
-        self.rate_limiter = pyrate_limiter.Limiter(bucket_factory, max_delay=math.inf)
-        
+                
         op_map = {op["name"]: op for op in self.config["operations"]}
 
         # Hash each pipeline step/operation
